+++ R/add_p.R
@@ -126,7 +126,11 @@ add_p.tbl_summary <- function(x,
 
   cards::process_formula_selectors(
     select_prep(x$table_body, x$inputs$data[include]),
-    test = test,
+    test =
+      case_switch(
+        missing(test) ~ get_theme_element("add_p.tbl_summary-arg:test", default = test),
+        .default = test
+      ),
     include_env = TRUE
   )
   # add the calling env to the test
@@ -143,7 +147,7 @@ add_p.tbl_summary <- function(x,
   # if `pvalue_fun` not modified, check if we need to use a theme p-value
   if (missing(pvalue_fun)) {
     pvalue_fun <-
-      get_theme_element("add_p.tbl_summary-arg:pvalue_fun") %||%
+      get_deprecated_theme_element("add_p.tbl_summary-arg:pvalue_fun") %||%
       get_theme_element("pkgwide-fn:pvalue_fun") %||%
       pvalue_fun
   }
@@ -448,7 +452,7 @@ calculate_and_add_test_results <- function(x, include, group, test.args, adj.var
     x <-
       modify_column_merge(
         x,
-        pattern = "{conf.low}, {conf.high}",
+        pattern = paste("{conf.low}", "{conf.high}", sep = get_theme_element("pkgwide-str:ci.sep", default = ", ")),
         rows = !is.na(.data$conf.low)
       )
   }