+++ R/utils-gtsummary_core.R
@@ -1,4 +1,3 @@
-get_theme_element <- function(..., default = NULL) default
 translate_text <- function(x, ...) x
 
 construct_initial_table_styling <- function(x) {